+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.7] - 2018-11-30
+## [0.2.0] - 2018-12-01
 ### Added
 
 ## [0.1.6] - 2018-11-30
@@ -79,7 +79,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `SQLiteDatabase#rawQuery()` method
 - `SQLiteDatabase#setLocale()` method
 
-[0.1.7]:  https://github.com/drydart/flutter_sqlcipher/compare/0.1.6...0.1.7
+[0.2.0]:  https://github.com/drydart/flutter_sqlcipher/compare/0.1.6...0.2.0
 [0.1.6]:  https://github.com/drydart/flutter_sqlcipher/compare/0.1.5...0.1.6
 [0.1.5]:  https://github.com/drydart/flutter_sqlcipher/compare/0.1.4...0.1.5
 [0.1.4]:  https://github.com/drydart/flutter_sqlcipher/compare/0.1.3...0.1.4